chore: move `installTemplate` out of `templates`
