+++ utilities/installTemplate.ts
